+++ Runtime/Basic/BasicState.cs
@@ -1,8 +1,0 @@
-﻿using UnityEngine;
-
-namespace GI.UnityToolkit.State
-{
-    [CreateAssetMenu(menuName = "State/Basic State/State")]
-    [System.Serializable]
-    public class BasicState : StateBase {}
-}