+++ Runtime/Basic/BasicStateListener.cs
@@ -1,4 +1,0 @@
-﻿namespace GI.UnityToolkit.State
-{
-    public class BasicStateListener : StateListenerBase<BasicState, BasicStateManager> {}
-}