+++ Runtime/Basic/BasicStateManager.cs
@@ -1,7 +1,0 @@
-﻿using UnityEngine;
-
-namespace GI.UnityToolkit.State
-{
-    [CreateAssetMenu(menuName = "State/Basic State/Manager")]
-    public class BasicStateManager : StateManagerBase<BasicState> {}
-}