+++ Runtime/Default/State.cs
@@ -1,0 +1,7 @@
+﻿using UnityEngine;
+
+namespace GI.UnityToolkit.State
+{
+    [CreateAssetMenu(menuName = "State/State")]
+    public class State : StateBase {}
+}