+++ Runtime/Default/StateListener.cs
@@ -1,0 +1,7 @@
+﻿using UnityEngine;
+
+namespace GI.UnityToolkit.State
+{
+    [AddComponentMenu("Grimbar Interactive/State/State Listener")]
+    public class StateListener : StateListenerBase<State, StateManager> {}
+}