+++ Runtime/Default/StateManager.cs
@@ -1,0 +1,7 @@
+﻿using UnityEngine;
+
+namespace GI.UnityToolkit.State
+{
+    [CreateAssetMenu(menuName = "State/Manager")]
+    public class StateManager : StateManagerBase<State> {}
+}