+++ Runtime/Default/StateTransitionListener.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+namespace GI.UnityToolkit.State
+{
+    [AddComponentMenu("Grimbar Interactive/State/State Transition Listener")]
+    public class StateTransitionListener : StateTransitionListenerBase<State, StateManager> {}
+}